requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.4
    # via black
attrs==21.2.0
    # via
    #   flake8-bugbear
    #   pytest
backcall==0.2.0
    # via ipython
black==21.7b0
    # via -r requirements-dev.in
click==8.0.1
    # via black
coverage==5.5
    # via pytest-cov
decorator==5.0.9
    # via ipython
factory-boy==3.2.0
    # via
    #   -r requirements-dev.in
    #   pytest-factoryboy
faker==8.11.0
    # via factory-boy
fastdiff==0.3.0
    # via snapshottest
flake8==3.9.2
    # via
    #   -r requirements-dev.in
    #   flake8-bugbear
    #   flake8-polyfill
    #   pep8-naming
flake8-bugbear==21.4.3
    # via -r requirements-dev.in
flake8-polyfill==1.0.2
    # via pep8-naming
freezegun==1.1.0
    # via -r requirements-dev.in
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
ipython==7.26.0
    # via -r requirements-dev.in
ipython-genutils==0.2.0
    # via traitlets
isort==5.9.3
    # via -r requirements-dev.in
jedi==0.18.0
    # via ipython
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==21.0
    # via
    #   -c requirements.txt
    #   pytest
parso==0.8.2
    # via jedi
pathspec==0.9.0
    # via black
pep8-naming==0.12.1
    # via -r requirements-dev.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.19
    # via ipython
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pygments==2.9.0
    # via ipython
pyparsing==2.4.7
    # via
    #   -c requirements.txt
    #   packaging
pytest==6.2.4
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
pytest-cov==2.12.1
    # via -r requirements-dev.in
pytest-django==4.4.0
    # via -r requirements-dev.in
pytest-factoryboy==2.1.0
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
regex==2021.8.3
    # via black
six==1.16.0
    # via
    #   -c requirements.txt
    #   python-dateutil
    #   snapshottest
snapshottest==0.6.0
    # via -r requirements-dev.in
termcolor==1.1.0
    # via snapshottest
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
tomli==1.2.1
    # via black
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
wasmer==1.0.0
    # via fastdiff
wasmer-compiler-cranelift==1.0.0
    # via fastdiff
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
